+++ src/main/java/ca/corefacility/bioinformatics/irida/service/sample/SampleService.java
@@ -1,5 +1,6 @@
 package ca.corefacility.bioinformatics.irida.service.sample;
 
+import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 
@@ -17,6 +18,7 @@ import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.sample.SampleSequencingObjectJoin;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequenceFile;
 import ca.corefacility.bioinformatics.irida.model.sequenceFile.SequencingObject;
+import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
 import ca.corefacility.bioinformatics.irida.service.CRUDService;
 
 /**
@@ -223,4 +225,14 @@ public interface SampleService extends CRUDService<Long, Sample> {
 	public Page<ProjectSampleJoin> getFilteredSamplesForProjects(List<Project> projects, List<String> sampleNames,
 			String sampleName, String searchTerm, String organism,
 			Date minDate, Date maxDate, int currentPage, int pageSize, Sort.Direction direction, String sortProperty);
+	
+	/**
+	 * Get a list of all {@link Sample}s associated with a given
+	 * {@link AnalysisSubmission}
+	 * 
+	 * @param submission
+	 *            the {@link AnalysisSubmission}
+	 * @return a Collection of {@link Sample}
+	 */
+	public Collection<Sample> getSamplesForAnalysisSubimssion(AnalysisSubmission submission);
 }