+++ src/main/webapp/resources/js/pages/projects/samples/project-samples.js
@@ -492,7 +492,7 @@
     function _selectAllSamples() {
       _sampleService.getAllIds()
         .then(function(result) {
-          $window.datatable.selectAll(result.data);
+          $window.datatable.selectAll(result);
         });
     }