+++ src/main/webapp/resources/js/pages/sequencing-runs/sequencing-runs-list.js
@@ -3,9 +3,19 @@ var runsTable = (function(page) {
 		return '<a class="item-link run-id" title="' + data + '" href="' + page.urls.link
 				+ full.id + '"><span>' + data + '</span></a>';		
 	};
+
+	function runUserRow(data, type, full) {
+		if(data != null){
+			return '<a class="item-link" title="' + data.label + '" href="' + page.urls.users + data.identifier + '"><span>' + data.label + '</span></a>';
+		}
+		else{
+			return "";
+		}	
+	};
 	
 	return {
-		runLinkRow : runLinkRow
+		runLinkRow : runLinkRow,
+		runUserRow : runUserRow
 
 	};
 })(window.PAGE);